[profiler] Add support to track range of fields in samples.
This adds support to track the range of values in a particular field
as samples are received. The lower and upper bounds of the field will
be stored in the data mapping and can be retrieved using:

  ppg_model_get_range(model, key, &lower, &upper)

To enable range tracking for a particular key, use:

  ppg_model_set_track_range(model, key, TRUE)

Range tracking is not currently supported on function based mappings
added with ppg_model_add_mapping_func().