[agent, libperfkit, profiler, shell] Add "started-at" out parameter to Channel::start
This allows clients to resolve timestamps from a given start offset
if needed. Primarily, this would be useful for the profiler as incoming
manifests may not be the actual time that the channel started.

Therefore, data models will be able to subtract the real start time from
the resolved sample timestamp.