[profiler] Binary search for the sample range.
This changeset will binary search for the samples that are found
within a given range. It aims to be inclusive, so it will include
a single sample on each side of the given range if possible. This
should make it easier for visualizers that need to draw to the edges.