[build] Lots of build system cleanup.
This goes through and cleans up how we do most of our building.
The old system was becomming a pain to maintain. This should be
a bit cleaner and easier to see how to modify for new contributors.
The Makefile.inc in the toplevel saves some redundancy as well.

Longer term, I'd like to take this a step further and get to a
non-recursive automake.